feat: 사용자 탈퇴 시 팀 리더일 경우 탈퇴 불가

+++ core/src/main/java/io/devarium/core/domain/user/service/UserServiceImpl.java
@@ -83,6 +83,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void withdraw(User user) {
+        if (teamService.checkUserIsLeader(user.getId())) {
+            throw new UserException(UserErrorCode.USER_WITHDRAW_FAILED, user.getId());
+        }
         user.delete();
         userRepository.save(user);
     }
